# Conflicts:
#	client/src/app/website/home/content/book-list/book-detail/book-detail.component.html

+++ client/src/app/admin/users/users.component.ts
@@ -1,6 +1,9 @@
 import { UserService } from './../../core/services/user.service';
 import { Component, OnInit } from '@angular/core';
 import { User } from '../../core';
+import swal from 'sweetalert2';
+import { Subject, of, Observable } from 'rxjs';
+import { debounceTime, distinctUntilChanged, switchMap, catchError, tap } from 'rxjs/operators';
 
 @Component({
   selector: 'admin-users',
@@ -8,30 +11,70 @@ import { User } from '../../core';
   styleUrls: ['./users.component.css']
 })
 export class AdminUsersComponent implements OnInit {
-
+  currentUser: User;
   users: User[] = new Array();
-  isEditor:boolean = false;
+  isEditor: boolean = false;
   selectedUser: User;
   buttonLabel: string = 'New User';
+  currentPage: number = 0;
+  userSize: number = 5; //get 5 items
+  totalPages: number = 0;
+  // pages: number[] = new Array();
+
+  private searchTerm$: Subject<string> = new Subject();
+
 
   constructor(
     private userService: UserService
   ) { }
 
   ngOnInit() {
+
+    this.loadUsers();
+    this.loadSearch();
+  }
+
+  loadSearch() {
+    this.search(this.searchTerm$).subscribe(
+      data => {
+        this.users = data.filter(data => (data.email != this.currentUser.email));
+      });
+
+  }
+  search(terms: Observable<string>) {
+    return terms.pipe(
+      debounceTime(400),
+      distinctUntilChanged(),
+      switchMap(term => this.userService.search(term))
+    )
+  }
+
+  loadUsers() {
+    let pageConfig = {
+      page: this.currentPage,
+      size: this.userSize
+    }
     this.userService.currentUser.subscribe(
       user => {
-        this.userService.getAll().subscribe(
+        this.currentUser = user;
+        this.userService.getAll(pageConfig).subscribe(
           data => {
-            this.users = data.filter(data => (data.email != user.email));
+            console.log(data);
+            this.users = data['content'];
+            this.totalPages = data['totalPages'];
           }
         )
       }
     )
   }
 
-  displayForm(user?: User){
-    if (user){
+  setPage(i) {
+    this.currentPage = i;
+    this.loadUsers();
+  }
+
+  displayForm(user?: User) {
+    if (user) {
       this.selectedUser = user;
     }
     this.isEditor = !this.isEditor;
@@ -40,32 +83,45 @@ export class AdminUsersComponent implements OnInit {
     console.log(this.isEditor);
   }
 
-  userEmit(user: User){
+  userEmit(user: User) {
     console.log(user);
-    if (user){
+    if (user) {
       this.users.push(user);
     }
     this.displayForm();
   }
 
-  delete(user: User){
-    this.userService.delete(user.id).subscribe(
-      data => {
-        this.users.splice(this.users.indexOf(user), 1)
-      }
-    )
-  }
-
-  searching(event){
-    this.userService.currentUser.subscribe(
-      user => {
-        this.userService.search(event.target.value).subscribe(
+  delete(user: User) {
+    swal({
+      title: 'Delete this user',
+      type: 'question',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      confirmButtonText: 'Yes'
+    }).then((result) => {
+      if (result.value) {
+        this.userService.delete(user.id).subscribe(
           data => {
-            this.users = data.filter(data => (data.email != user.email));
+            this.users.splice(this.users.indexOf(user), 1);
+            swal({
+              type: 'success',
+              title: 'Successful',
+            })
           }
         )
       }
-    )
+    })
+
+  }
+
+  searching(event) {
+    console.log('keyup: ' + event.target.value)
+    if (event.target.value == ''){
+      this.loadUsers();
+    } else {
+      this.searchTerm$.next(event.target.value);
+    }
+    
   }
